Move some files to stlite-kernel/src/streamlit-replacements/lib
