+++ packages/stlite-kernel/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./kernel"
-export * from "./ConnectionManager"
-export * from "./ConnectionState"
+export * from "./streamlit-replacements/lib/ConnectionManager"
+export * from "./streamlit-replacements/lib/ConnectionState"
 export * from "./react-helpers"
