+++ packages/stlite-kernel/src/streamlit-replacements/lib/ConnectionManager.ts
@@ -3,7 +3,7 @@
 import { BackMsg, ForwardMsg } from "streamlit-browser/src/autogen/proto"
 import { ReactNode } from "react"
 
-import { StliteKernel } from "./kernel"
+import { StliteKernel } from "../../kernel"
 import { ConnectionState } from "./ConnectionState"
 import { ensureError } from "./ErrorHandling"
