+++ packages/stlite-kernel/src/streamlit-replacements/lib/ConnectionState.ts
